Add missing retry logic in Celery tasks
Ensured all tasks are idempotent for retry.
Was going to DRY up the retry logic, but realized that we don't have
Job records in the db for all task types, and we're ending up updating
the status of the parent generate_media_task instead, which is not ideal.
So we'd probably want to create Job records for all tasks before we
refactor this.